fix: user can only leave if he/she did not start the event

+++ app/event/page.tsx
@@ -85,6 +85,11 @@ export default function Event() {
     }
 
     async function leave(userId: string, eventId: string) {
+        // User is the one who starts the event.
+        if (eventDetails?.email ?? null === email) {
+            return;
+        }
+
         const { error: error1 } = await supabase
             .from('participant')
             .delete()
